refactor: window controls and centralize layout constants
Extracted window control buttons into a reusable WindowControls component and replaced hardcoded layout values with centralized APP_SHELL constants. Updated all relevant components and pages to use APP_SHELL for header height, icon size, and navbar width, improving maintainability and consistency.
